refactor: make TaskState as a class
